Decouple mobile-web submodule
For other OAuth2 providers to extend easily, in addition to just support Google Accounts.
