+++ auth-api-gms/build.gradle.kts
@@ -7,7 +7,7 @@ android {
 }
 
 dependencies {
-    api("com.openmobilehub.android:auth-api:1.0.1-beta")
+    api(project(":auth-api"))
 
     // KTX
     implementation(Libs.coreKtx)
