+++ auth-api-mobileweb/gradle.properties
@@ -1,0 +1,4 @@
+artifactId=auth-non-gms
+version=1.0.1-beta
+googleAuthUrl="https://oauth2.googleapis.com"
+description=The non GMS implementation of the OMH Auth API
