+++ auth-api-mobileweb/src/main/java/com/omh/android/auth/mobileweb/data/login/models/AuthTokenResponse.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2023 Open Mobile Hub
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.omh.android.auth.mobileweb.data.login.models
+
+import androidx.annotation.Keep
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import com.fasterxml.jackson.annotation.JsonProperty
+
+@Keep
+@JsonIgnoreProperties(ignoreUnknown = true)
+class AuthTokenResponse(
+    @JsonProperty("access_token")
+    val accessToken: String,
+    @JsonProperty("token_type")
+    val tokenType: String,
+    @JsonProperty("expires_in")
+    val expiresIn: Long,
+    @JsonProperty("refresh_token")
+    val refreshToken: String?,
+    @JsonProperty("id_token")
+    val idToken: String
+)
