+++ auth-api-mobileweb/src/main/java/com/omh/android/auth/mobileweb/domain/user/ProfileUseCase.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2023 Open Mobile Hub
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.omh.android.auth.mobileweb.domain.user
+
+import com.omh.android.auth.api.models.OmhAuthException
+import com.omh.android.auth.api.models.OmhUserProfile
+
+class ProfileUseCase(private val userRepository: UserRepository) {
+
+    @Throws(OmhAuthException.UnrecoverableLoginException::class)
+    suspend fun resolveIdToken(idToken: String, clientId: String) {
+        if (idToken.trim().isEmpty() || clientId.trim().isEmpty()) {
+            throw OmhAuthException.UnrecoverableLoginException(
+                cause = IllegalStateException("idToken or clientId is empty")
+            )
+        }
+        userRepository.handleIdToken(idToken, clientId)
+    }
+
+    fun getProfileData(): OmhUserProfile? {
+        return userRepository.getProfileData()
+    }
+
+    companion object {
+
+        fun createUserProfileUseCase(userRepository: UserRepository): ProfileUseCase {
+            return ProfileUseCase(userRepository)
+        }
+    }
+}
