+++ auth-api-non-gms/src/main/java/com/omh/android/auth/nongms/data/utils/GoogleRetrofitImpl.kt
@@ -16,9 +16,9 @@
 
 package com.omh.android.auth.nongms.data.utils
 
+import com.omh.android.auth.mobileweb.data.utils.retrofit.ApiResultCallAdapterFactory
 import com.omh.android.auth.nongms.BuildConfig
 import com.omh.android.auth.nongms.data.login.GoogleAuthREST
-import com.omh.android.auth.nongms.data.utils.retrofit.ApiResultCallAdapterFactory
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
