+++ auth-api-non-gms/src/test/java/com/omh/android/auth/nongms/repositories/UserRepositoryTest.kt
@@ -17,9 +17,9 @@
 package com.omh.android.auth.nongms.repositories
 
 import com.omh.android.auth.api.models.OmhUserProfile
+import com.omh.android.auth.mobileweb.data.user.datasource.UserDataSource
+import com.omh.android.auth.mobileweb.domain.user.UserRepository
 import com.omh.android.auth.nongms.data.user.UserRepositoryImpl
-import com.omh.android.auth.nongms.data.user.datasource.UserDataSource
-import com.omh.android.auth.nongms.domain.user.UserRepository
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
